Implemented 10s PIPR
Following the design of Kankipati 2010. 10 second PIPR pulses (with 0.5 half-cosine ramps on and off, 9 sec plateau), with a 10 second pre-stimulus interval and an ~60s post-stimulus interval.

We have not yet set up the long pulse version of the LMS and Mel silent substitution stimuli, but it seems that this is inveitable.

Signed-off-by: gkaguirre <gkaguirre@me.com>